Odin Project Exercise: Landing Page
A Website that uses flexbox to adjust items positions.

There is no reason for the choice of the topic but
other than that I am reading the books at the moment.

I own no rights or any relation to the links at this website.
Check George R. R. Martin website/blog for further information.